fix: address Round 1 review — deduplicate SAFE_REASON_CODE, move IdempotencyKeyUtilTest to tramai-core, add KDoc to ToolExecutionContext.idempotencyKey
