+++ tramai-security/src/main/kotlin/dev/tramai/security/approval/InMemoryApprovalRecoveryCoordinator.kt
@@ -6,6 +6,7 @@ import dev.tramai.core.approval.ApprovalLifecycleAuditEmitter
 import dev.tramai.core.approval.ApprovalRecoveryCoordinator
 import dev.tramai.core.approval.ForceCancelClaimedCommand
 import dev.tramai.core.approval.NoOpApprovalLifecycleAuditEmitter
+import dev.tramai.core.approval.SAFE_REASON_CODE_PATTERN
 import dev.tramai.core.exception.ApprovalAuthorizationException
 import dev.tramai.core.exception.ApprovalContinuationConflictException
 import dev.tramai.core.exception.ApprovalContinuationNotFoundException
@@ -25,7 +26,7 @@ class InMemoryApprovalRecoveryCoordinator(
     private val clock: Clock = Clock.systemUTC(),
 ) : ApprovalRecoveryCoordinator {
 
-    private val SAFE_REASON_CODE = Regex("[a-z0-9][a-z0-9._:-]{0,63}")
+    private val SAFE_REASON_CODE = Regex(SAFE_REASON_CODE_PATTERN)
 
     override suspend fun findStaleClaims(
         claimedBefore: Instant,
